+++ src/services/stores/queries/businessApplication.js
@@ -502,6 +502,7 @@ query _getBusinessApplicationById ($id: String!) {
   }
 }
 `;
+<<<<<<< HEAD
 
 export const signPortalAgreement = gql`
 mutation _signPortalAgreement($applicationId: String!, $issuerId: String!, $selectedOffer: OfferInput!, $isSelectedOfferChanged: Boolean){
@@ -532,3 +533,5 @@ mutation _createOffering($applicationId: String!){
   }
 }
 `;
+=======
+>>>>>>> develop